# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Коробыш API
  version: 1.0.0
  description: |
    Контракт API для персональных наград, ежемесячного рейтинга, еженедельных
    наград за вход, ежедневных заданий и ежедневной сводки.

    Правила работы со временем:
    - все временные метки передаются в формате RFC3339 и часовом поясе UTC;
    - периоды рейтинга соответствуют календарным месяцам в UTC;
    - циклы еженедельных входов соответствуют неделям ISO с понедельника по воскресенье в UTC;
    - сервер является источником истины для идентификатора игрока, дат, размеров наград,
      уровней, заработанных листьев, позиций и статуса проверки.
servers:
  - url: /
    description: Бэкенд приложения
...
security:
  - bearerAuth: []
paths:
...
  /api/app/auth/request-otp:
    post:
      tags: [Авторизация]
      operationId: requestOtp
      summary: Запросить одноразовый код для входа
      description: |
        Нормализует адрес электронной почты, создаёт пользователя при первом обращении
        и отправляет новый одноразовый код. Ранее выданный код для этого пользователя
        становится недействительным.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
            examples:
              default:
                value:
                  email: user@example.com
      responses:
        '200':
          description: Одноразовый код успешно отправлен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpSentResponse'
              examples:
                default:
                  value:
                    sent: true
        '400':
          description: Тело запроса некорректно или содержит недопустимый адрес электронной почты.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: Не удалось создать или отправить одноразовый код.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/app/auth/verify-otp:
    post:
      tags: [Авторизация]
      operationId: verifyOtp
      summary: Подтвердить одноразовый код
      description: |
        Проверяет одноразовый код и возвращает JWT вместе с данными пользователя.
        Код можно ввести неверно не более трёх раз; просроченный или успешно
        использованный код становится недействительным.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerificationRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  code: '12345678'
      responses:
        '200':
          description: Код подтверждён, JWT выпущен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                default:
                  value:
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example
                    record:
                      id: 72482c10-6976-4823-a835-333333333333
                      email: user@example.com
                      verified: true
        '400':
          description: Тело запроса некорректно либо код недействителен или просрочен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: Не удалось проверить код или выпустить JWT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/app/auth/me:
    get:
      tags: [Авторизация]
      operationId: getCurrentUser
      summary: Получить текущего пользователя
      description: Определяет пользователя по JWT из заголовка `Authorization`.
      responses:
        '200':
          description: Данные авторизованного пользователя.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                default:
                  value:
                    id: 72482c10-6976-4823-a835-333333333333
                    email: user@example.com
                    verified: true
        '401':
          description: JWT отсутствует, недействителен или принадлежит неизвестному пользователю.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: Не удалось загрузить пользователя.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/v1/pet:
    get:
      tags: [Питомец]
      operationId: getPet
      summary: Получить питомца текущего пользователя
      description: |
        Возвращает единственного питомца, связанного с авторизованным пользователем.
        Уровень вычисляется сервером по оставшемуся балансу листьев и не принимается
        от клиента. Листья тратятся на переходы между уровнями, поэтому баланс уменьшается
        после каждого повышения. Новый пользователь начинает с питомцем 1-го уровня и
        пустым именем.
        Если `name` пустое, клиент показывает пользователю форму первичной настройки имени.
      responses:
        '200':
          description: Текущее состояние питомца.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
              examples:
                default:
                  value:
                    name: Листик
                    level: 5
                    leaves: 45
                uninitialized:
                  value:
                    name: ''
                    level: 1
                    leaves: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Питомец]
      operationId: updatePetName
      summary: Изменить имя питомца
      description: Обновляет имя питомца текущего пользователя. Уровень и листья этим запросом не изменяются.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetNameUpdateRequest'
            examples:
              default:
                value:
                  name: Листик
      responses:
        '200':
          description: Имя успешно изменено.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '400':
          description: Имя отсутствует, пустое или превышает допустимую длину.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PetNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/pet/ws:
    get:
      tags: [Питомец]
      operationId: subscribePetProgress
      summary: Подписаться на прогресс питомца
      description: |
        WebSocket-подключение для авторизованного пользователя. После подключения сервер
        сразу отправляет полный снимок прогресса и повторяет его после каждого изменения
        баланса листьев или уровня питомца. Листья после повышения сгорают, а сервер
        пересчитывает уровень по оставшемуся балансу. Поле `levelUp` сообщает, повысился
        ли уровень именно в операции, вызвавшей это сообщение.

        Для подключения используется тот же JWT Bearer-токен, что и для HTTP API.
        Браузерные клиенты могут передать его в query-параметре `token`, поскольку
        стандартный WebSocket API не позволяет задать Authorization-заголовок.
        В OpenAPI 3.0.3 WebSocket описан расширением `x-websocket`.
      x-websocket:
        serverMessages:
          event: PET_PROGRESS_UPDATED
          schema:
            $ref: '#/components/schemas/PetProgressUpdatedEvent'
      responses:
        '101':
          description: Соединение переключено на протокол WebSocket.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/pet/levels:
    get:
      tags: [Питомец]
      operationId: getPetLevels
      summary: Получить уровни и награды питомца
      description: |
        Возвращает все десять уровней в порядке возрастания. У каждого уровня есть
        награда и её состояние. Когда пользователь достигает уровня, награда получает
        статус `UNOPENED` и должна быть забрана в течение трёх суток.

        Для `UNOPENED` поле `expiresAt` является дедлайном, по которому клиент отображает
        обратный отсчёт. После его наступления сервер переводит награду в `FROZEN`; для
        этого статуса в поле остаётся прошедшая метка дедлайна. Для `LOCKED` и `CLAIMED`
        значение равно null.
      responses:
        '200':
          description: Полная шкала уровней и состояние награды каждого уровня.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetLevelsResponse'
              examples:
                petAtLevelFour:
                  value:
                    levels:
                      - level: 1
                        status: CLAIMED
                        reward:
                          id: 10101010-1010-4010-8010-101010101010
                          type: AVITO_BONUS
                          description: 50 бонусов Авито
                        expiresAt: null
                      - level: 2
                        status: CLAIMED
                        reward:
                          id: 20202020-2020-4020-8020-202020202020
                          type: FREE_DELIVERY
                          description: Бесплатная доставка для одного заказа
                        expiresAt: null
                      - level: 3
                        status: FROZEN
                        reward:
                          id: 30303030-3030-4030-8030-303030303030
                          type: PROMOTION_DISCOUNT
                          description: Скидка 10% на продвижение объявления
                        expiresAt: '2026-08-03T12:00:00Z'
                      - level: 4
                        status: UNOPENED
                        reward:
                          id: 40404040-4040-4040-8040-404040404040
                          type: FREE_PROMOTION
                          description: Бесплатное продвижение объявления на 3 дня
                        expiresAt: '2026-08-09T12:00:00Z'
                      - level: 5
                        status: LOCKED
                        reward:
                          id: 50505050-5050-4050-8050-505050505050
                          type: DELIVERY_DISCOUNT
                          description: Скидка 15% на Авито Доставку
                        expiresAt: null
                      - level: 6
                        status: LOCKED
                        reward:
                          id: 60606060-6060-4060-8060-606060606060
                          type: AVITO_BONUS
                          description: 100 бонусов Авито
                        expiresAt: null
                      - level: 7
                        status: LOCKED
                        reward:
                          id: 70707070-7070-4070-8070-707070707070
                          type: FREE_DELIVERY
                          description: Бесплатная доставка для двух заказов
                        expiresAt: null
                      - level: 8
                        status: LOCKED
                        reward:
                          id: 80808080-8080-4080-8080-808080808080
                          type: PROMOTION_DISCOUNT
                          description: Скидка 20% на продвижение объявления
                        expiresAt: null
                      - level: 9
                        status: LOCKED
                        reward:
                          id: 90909090-9090-4090-8090-909090909090
                          type: FREE_PROMOTION
                          description: Бесплатное продвижение объявления на 7 дней
                        expiresAt: null
                      - level: 10
                        status: LOCKED
                        reward:
                          id: a0a0a0a0-a0a0-40a0-80a0-a0a0a0a0a0a0
                          type: DELIVERY_DISCOUNT
                          description: Скидка 25% на Авито Доставку
                        expiresAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PetNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/pet/level-rewards/{rewardId}/claim:
    post:
      tags: [Питомец]
      operationId: claimPetLevelReward
      summary: Забрать награду за уровень
      description: |
        Переводит указанную награду уровня из `UNOPENED` в `CLAIMED`. Награды остальных
        уровней этим запросом не изменяются. Сервер сопоставляет `rewardId` с
        авторизованным пользователем и уровнем, после чего добавляет награду с источником
        `LEVEL` в одной транзакции.
      parameters:
        - name: rewardId
          in: path
          required: true
          description: Идентификатор награды из соответствующего элемента `levels`.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Награда за указанный уровень успешно забрана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimLevelRewardResponse'
              examples:
                claimed:
                  value:
                    level: 4
                    status: CLAIMED
        '400':
          description: Передан некорректный идентификатор награды.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Питомец не найден, либо награда не принадлежит авторизованному пользователю.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Награду нельзя забрать в её текущем состоянии.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                locked:
                  value:
                    code: LEVEL_REWARD_LOCKED
                    message: Сначала необходимо достичь этого уровня
                frozen:
                  value:
                    code: LEVEL_REWARD_FROZEN
                    message: Срок получения награды истёк
                claimed:
                  value:
                    code: LEVEL_REWARD_ALREADY_CLAIMED
                    message: Награда за этот уровень уже забрана
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/pet/chests/open:
    post:
      tags: [Питомец]
      operationId: openPetChest
      summary: Открыть сундук за листья
      description: |
        Открывает один сундук для авторизованного пользователя. Операция доступна
        только после достижения 10-го уровня и стоит ровно 200 листьев.

        Сервер в одной транзакции проверяет уровень и баланс, списывает листья,
        создаёт запись об открытии и сразу выдаёт одну существенную награду уровня
        10 с источником `CHEST`. Выданная награда немедленно доступна в разделе
        «Награды»; отдельного подтверждения получения не требуется.
      responses:
        '200':
          description: Сундук открыт, листья списаны и награда выдана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardItem'
              examples:
                opened:
                  value:
                    id: c2323232-2323-4232-8232-232323232323
                    title: Бесплатная доставка для трёх заказов
                    category: FREE_DELIVERY
                    categoryName: Бесплатная доставка
                    source: CHEST
                    active: true
                    status: ACTIVE
                    expiresAt: '2026-11-07T12:00:00Z'
                    awardedAt: '2026-08-07T12:00:00Z'
                    redeemedAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Питомец пользователя ещё не создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                petNotFound:
                  value:
                    code: PET_NOT_FOUND
                    message: Питомец пользователя не найден.
        '409':
          description: Сундук недоступен до 10-го уровня либо на балансе меньше 200 листьев.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                levelRequired:
                  value:
                    code: CHEST_LEVEL_REQUIRED
                    message: Сундуки доступны с 10-го уровня питомца.
                insufficientLeaves:
                  value:
                    code: INSUFFICIENT_LEAVES
                    message: Для открытия сундука требуется 200 листьев.
        '500':
          $ref: '#/components/responses/InternalError'
  /api/app/rewards:
    get:
      tags: [Награды]
      operationId: getRewards
      summary: Получить все награды
      description: |
        Возвращает полную историю персональных наград пользователя, включая
        активные, использованные и просроченные. Награды сгруппированы по категориям,
        внутри группы сначала отображаются выданные позднее.
      responses:
        '200':
          description: Все награды пользователя, сгруппированные по категориям.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardsResponse'
              examples:
                default:
                  value:
                    groups:
                      - category: AVITO_BONUS
                        categoryName: Бонусы Авито
                        items:
                          - id: 49a70fa1-4a2a-4cc7-b09d-444444444444
                            title: 100 бонусов
                            category: AVITO_BONUS
                            categoryName: Бонусы Авито
                            source: LEVEL
                            active: true
                            status: ACTIVE
                            expiresAt: '2026-09-04T12:00:00Z'
                            awardedAt: '2026-08-05T12:00:00Z'
                            redeemedAt: null
                      - category: FREE_DELIVERY
                        categoryName: Бесплатная доставка
                        items:
                          - id: 2c9f7f9f-a958-4e77-b17c-555555555555
                            title: Бесплатная доставка (1 заказ)
                            category: FREE_DELIVERY
                            categoryName: Бесплатная доставка
                            source: CHEST
                            active: false
                            status: REDEEMED
                            expiresAt: '2026-08-20T12:00:00Z'
                            awardedAt: '2026-08-05T12:00:00Z'
                            redeemedAt: '2026-08-05T13:30:00Z'
        '401':
          description: JWT отсутствует или недействителен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: Не удалось загрузить награды.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/app/rewards/{rewardId}:
    get:
      tags: [Награды]
      operationId: getReward
      summary: Открыть награду
      description: |
        Возвращает награду текущего пользователя по идентификатору, включая
        использованные и просроченные награды.
      parameters:
        - name: rewardId
          in: path
          required: true
          description: Идентификатор награды.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Полные данные награды и её текущее состояние.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardDetails'
        '400':
          description: Передан некорректный идентификатор награды.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: JWT отсутствует или недействителен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: Награда не существует или принадлежит другому пользователю.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: Не удалось загрузить награду.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/app/rewards/{rewardId}/redeem:
    post:
      tags: [Награды]
      operationId: redeemReward
      summary: Использовать награду
      description: |
        Помечает активную награду текущего пользователя как использованную.
        Награда остаётся в общем списке со статусом `REDEEMED` и `active: false`.
      parameters:
        - name: rewardId
          in: path
          required: true
          description: Идентификатор награды.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Награда успешно использована.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedeemRewardResponse'
              examples:
                default:
                  value:
                    id: 49a70fa1-4a2a-4cc7-b09d-444444444444
                    redeemedAt: '2026-08-05T13:30:00Z'
        '400':
          description: Передан некорректный идентификатор награды.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: JWT отсутствует или недействителен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: Активная награда не найдена, уже использована или просрочена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: Не удалось использовать награду.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/v1/tasks:
    get:
      tags: [Ежедневные задания]
      operationId: getDailyTasks
      summary: Получить задания на текущий день
      description: |
        Возвращает полный список из четырёх заданий текущего дня.
        Задания, недоступные на текущем уровне, также входят в список со статусом `LOCKED`.
        Прогресс обновляется на основании подтверждённых действий пользователя.
        Содержание заданий выбирается из серверной конфигурации отдельно для каждого UTC-дня.
        Награда и уровень открытия зависят от слота и не меняются: слоты 1 и 2 дают по 45 листьев
        и открыты с 1 уровня, слот 3 даёт 50 листьев и открывается с 5 уровня, слот 4 даёт 60
        листьев и открывается с 10 уровня.
        Уровень питомца сервер получает из состояния авторизованного пользователя;
        клиент не может передать или переопределить его через query-параметры.
      responses:
        '200':
          description: Задания и их текущий прогресс.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyTasksResponse'
              examples:
                default:
                  value:
                    tasks:
                      - taskId: 8f123af2-24c5-4ff7-b114-111111111111
                        slot: 1
                        type: VIEW_LISTINGS
                        description: Посмотреть 5 объявлений
                        currentCount: 5
                        targetCount: 5
                        rewardLeaves: 45
                        requiredLevel: 1
                        status: CLAIMED
                      - taskId: 6ab41fc5-1a5f-4db4-9129-222222222222
                        slot: 2
                        type: ADD_TO_FAVORITES
                        description: Добавить 3 объявления в избранное
                        currentCount: 3
                        targetCount: 3
                        rewardLeaves: 45
                        requiredLevel: 1
                        status: COMPLETED
                      - taskId: 72482c10-6976-4823-a835-333333333333
                        slot: 3
                        type: PUBLISH_LISTING
                        description: Опубликовать объявление
                        currentCount: 0
                        targetCount: 1
                        rewardLeaves: 50
                        requiredLevel: 5
                        status: LOCKED
                      - taskId: b826c297-bc5f-4e1a-93d5-444444444444
                        slot: 4
                        type: COMPLETE_DEAL
                        description: Совершить покупку или продажу
                        currentCount: 0
                        targetCount: 1
                        rewardLeaves: 60
                        requiredLevel: 10
                        status: LOCKED
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tasks/progress:
    get:
      tags: [Ежедневные задания]
      operationId: getDailyTasksProgress
      summary: Получить количество выполненных заданий за день
      description: |
        Возвращает компактный прогресс для отображения на страницах приложения.
        `completedCount` учитывает задания со статусами `COMPLETED` и `CLAIMED`.
        В `totalCount` входят все четыре задания, включая недоступные по уровню.
        Уровень питомца сервер получает из состояния авторизованного пользователя;
        клиент не может передать или переопределить его через query-параметры.
      responses:
        '200':
          description: Количество выполненных заданий за текущий день.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyTasksProgressResponse'
              examples:
                default:
                  value:
                    completedCount: 2
                    totalCount: 4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    OtpRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Адрес электронной почты для получения одноразового кода.
          example: user@example.com
    OtpVerificationRequest:
      type: object
      additionalProperties: false
      required: [email, code]
      properties:
        email:
          type: string
          format: email
          description: Адрес электронной почты, на который был отправлен код.
          example: user@example.com
        code:
          type: string
          pattern: '^\d{6,10}$'
          minLength: 6
          maxLength: 10
          description: Цифровой одноразовый код. Длина задаётся конфигурацией сервера, по умолчанию — 8 цифр.
          example: '12345678'
    OtpSentResponse:
      type: object
      required: [sent]
      properties:
        sent:
          type: boolean
          enum: [true]
          description: Подтверждает успешную отправку одноразового кода.
...
    PetNameUpdateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 35
          description: Новое имя питомца без начальных и конечных пробелов.
...
    AuthResponse:
      type: object
      required: [token, record]
      properties:
        token:
          type: string
          description: JWT для передачи в заголовке `Authorization` последующих запросов.
        record:
          $ref: '#/components/schemas/UserResponse'
    AuthErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Текст ошибки.
...